+++ lib/libtpe/tpe/task/ont_block_compression_task.cpp
@@ -29,7 +29,7 @@ void OntBlockCompressionTask::init(
     PBlockBuffer_t block_buffer,
     PTransactionService_t tx_service,
     bool enable_block_compression,
-    float min_tx_age_seconds
+    double min_tx_age_seconds
 )
 {
     _tx_service = std::move(tx_service);
